🔨 Refactor: 모달 폴더 생성
