+++ src/pages/ListPage/Credit/components/CreditModal/styles.module.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/';
+@import '../../../../../styles/';
 
 .modal {
   width: 295px;
