+++ src/pages/ListPage/MonthlyChart/components/ChartModal/styles.module.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/';
+@import '../../../../../styles/';
 
 .modal {
   width: 493px;
